Refactor data loading code into separate file
